fix(dropdown): correctly find index of active option in a hash
When the option list is a hash, a error was thrown when finding the active option
element because .indexOf() was used, method that is only available for arrays.
Now, we iterate over the object keys, trying to find which of them has the value
of the active option, then use that as the index for the option element.