fix(frontkit.less): don't hard reference normalize.css
We were hard referencing normalize.css in out frontkit.less file.
This was causing 3rd party code to break when importing it, because the
file would not exist in their setup.
Now, we'll prepend it to the frontkit*.css files via Grunt, instead of doing
so in the LESS compiler.
Code which is not using an pre-built frontkit.css must reference
normalize.css another way.

Fixes #57